added user check feature

+++ src/main/java/br/com/faunora/services/ExameService.java
@@ -2,21 +2,27 @@ package br.com.faunora.services;
 
 import br.com.faunora.domain.dto.ExameRecordDto;
 import br.com.faunora.domain.enums.ExameTipo;
+import br.com.faunora.domain.enums.UserTipo;
 import br.com.faunora.domain.models.ExameModel;
 import br.com.faunora.domain.models.PetModel;
+import br.com.faunora.domain.models.UserModel;
 import br.com.faunora.infra.exceptions.ExameNaoEncontradoException;
 import br.com.faunora.infra.exceptions.NenhumExameEncontradoException;
 import br.com.faunora.infra.exceptions.PetNaoEncontradoException;
+import br.com.faunora.infra.exceptions.UsuarioNaoEncontradoException;
 import br.com.faunora.repositories.ExameRepository;
 import br.com.faunora.repositories.PetRepository;
+import br.com.faunora.repositories.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.chrono.ChronoLocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 import java.util.UUID;
 
 @Service
@@ -25,11 +31,24 @@ public class ExameService {
     private ExameRepository exameRepository;
     @Autowired
     private PetRepository petRepository;
+    @Autowired
+    private UserRepository userRepository;
 
     @Transactional
     public void saveExame(ExameRecordDto exameRecordDto) {
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
         PetModel petModel = petRepository.findById(exameRecordDto.pacienteId())
                 .orElseThrow(PetNaoEncontradoException::new);
+
+        if (!userModel.getPets().contains(petModel)) {
+            throw new PetNaoEncontradoException();
+        }
 
         ExameModel exameModel = new ExameModel();
         exameModel.setPaciente(petModel);
@@ -37,16 +56,44 @@ public class ExameService {
         exameModel.setData(exameRecordDto.data());
         exameModel.setHora(exameRecordDto.hora());
 
+        List<UserModel> veterinarios = userRepository.findAllByTipo(UserTipo.VETERINÁRIO);
+
+        exameModel.setVeterinario(veterinarios.get(new Random().nextInt(veterinarios.size())));
+
         exameRepository.save(exameModel);
     }
 
     public ExameModel findById(Long id) {
-        return exameRepository.findById(id)
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
+        ExameModel exameModel = exameRepository.findById(id)
                 .orElseThrow(ExameNaoEncontradoException::new);
+
+        if (!userModel.getPets().contains(exameModel.getPaciente())) {
+            throw new PetNaoEncontradoException();
+        }
+
+        return exameModel;
     }
 
     public List<ExameModel> findAll() {
-        List<ExameModel> exameModels = exameRepository.findAll();
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
+        List<ExameModel> exameModels = new ArrayList<>();
+
+        for (PetModel petModel : userModel.getPets()) {
+            exameModels.addAll(exameRepository.findAllByPaciente(petModel));
+        }
 
         if (exameModels.isEmpty()) {
             throw new NenhumExameEncontradoException();
@@ -56,8 +103,19 @@ public class ExameService {
     }
 
     public List<ExameModel> findAllByPaciente(UUID pacienteId) {
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
         PetModel paciente = petRepository.findById(pacienteId)
                 .orElseThrow(PetNaoEncontradoException::new);
+
+        if (!userModel.getPets().contains(paciente)) {
+            throw new PetNaoEncontradoException();
+        }
 
         List<ExameModel> exameModels = exameRepository.findAllByPaciente(paciente);
 
@@ -69,7 +127,17 @@ public class ExameService {
     }
 
     public List<ExameModel> findAllByTipo(ExameTipo tipo) {
-        List<ExameModel> exameModels = exameRepository.findAllByTipo(tipo);
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
+        List<ExameModel> exameModels = new ArrayList<>();
+        for (PetModel petModel : userModel.getPets()) {
+            exameModels.addAll(exameRepository.findAllByTipoAndPaciente(tipo, petModel));
+        }
 
         if (exameModels.isEmpty()) {
             throw new NenhumExameEncontradoException();
@@ -79,7 +147,7 @@ public class ExameService {
     }
 
     public List<ExameModel> findAnteriores() {
-        List<ExameModel> exameModels = exameRepository.findAll();
+        List<ExameModel> exameModels = this.findAll();
 
         exameModels.removeIf(exameModel -> exameModel.getData().isAfter(ChronoLocalDate.from(LocalDate.now())));
 
@@ -91,7 +159,7 @@ public class ExameService {
     }
 
     public List<ExameModel> findMarcados() {
-        List<ExameModel> exameModels = exameRepository.findAll();
+        List<ExameModel> exameModels = this.findAll();
 
         exameModels.removeIf(exameModel -> exameModel.getData().isBefore(ChronoLocalDate.from(LocalDate.now())));
 
@@ -105,7 +173,7 @@ public class ExameService {
     public List<ExameModel> findAllByRandom(String filter) {
         List<ExameModel> exameModelsByRandom = new ArrayList<>();
 
-        for (ExameModel exameModel : exameRepository.findAll()) {
+        for (ExameModel exameModel : this.findAll()) {
             if (exameModel.getTipo().toString().contains(filter) || exameModel.getData().toString().contains(filter) || exameModel.getHora().toString().contains(filter)) {
                 exameModelsByRandom.add(exameModel);
             }
@@ -120,11 +188,22 @@ public class ExameService {
 
     @Transactional
     public void updateExame(Long id, ExameRecordDto exameRecordDto) {
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
         ExameModel exameModel = exameRepository.findById(id)
                 .orElseThrow(ExameNaoEncontradoException::new);
 
         PetModel petModel = petRepository.findById(exameRecordDto.pacienteId())
                 .orElseThrow(PetNaoEncontradoException::new);
+
+        if (!userModel.getPets().contains(petModel)) {
+            throw new PetNaoEncontradoException();
+        }
 
         exameModel.setPaciente(petModel);
         exameModel.setTipo(exameRecordDto.tipo());
@@ -136,8 +215,19 @@ public class ExameService {
 
     @Transactional
     public void deleteById(Long id) {
+        if (SecurityContextHolder.getContext().getAuthentication() == null) {
+            throw new UsuarioNaoEncontradoException();
+        }
+
+        UserModel userModel = userRepository.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName())
+                .orElseThrow(UsuarioNaoEncontradoException::new);
+
         ExameModel exameModel = exameRepository.findById(id)
                 .orElseThrow(ExameNaoEncontradoException::new);
+
+        if (!userModel.getPets().contains(exameModel.getPaciente())) {
+            throw new PetNaoEncontradoException();
+        }
 
         exameRepository.delete(exameModel);
     }
